refactor: 스웨거용 인터페이스 분리
