+++ src/main/java/com/example/happyDream/Interface/ChargerStateSwagger.java
@@ -1,7 +1,7 @@
-package com.example.happyDream.RestController;
+package com.example.happyDream.Interface;
 
 import io.swagger.v3.oas.annotations.tags.Tag;
 
 @Tag(name = "ChargerState", description = "충전기 상태 API")
-public interface ChargerState {
+public interface ChargerStateSwagger {
 }
