+++ src/main/java/com/example/happyDream/RestController/ChargerLogSwaggerRestController.java
@@ -3,18 +3,9 @@ package com.example.happyDream.RestController;
 import com.example.happyDream.DTO.ChargerDTO;
 import com.example.happyDream.DTO.ChargerLogDTO;
 import com.example.happyDream.DTO.ResponseDTO;
+import com.example.happyDream.Interface.ChargerLogSwagger;
 import com.example.happyDream.Service.ChargerLogService;
 import com.example.happyDream.Service.ChargerServiceFacade;
-import com.example.happyDream.Util.GsonUtil;
-import com.google.gson.Gson;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.enums.ParameterIn;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.ExampleObject;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
@@ -27,11 +18,11 @@ import java.util.List;
 @Slf4j
 @RestController
 @RequestMapping(path = "/api/v1", produces = "application/json")
-public class ChargerLogRestController implements ChargerLog{
+public class ChargerLogSwaggerRestController implements ChargerLogSwagger {
     private final ChargerServiceFacade chargerServiceFacade;
 
     @Autowired
-    public ChargerLogRestController(ChargerServiceFacade chargerServiceFacade, ChargerLogService chargerLogService) {
+    public ChargerLogSwaggerRestController(ChargerServiceFacade chargerServiceFacade, ChargerLogService chargerLogService) {
         this.chargerServiceFacade = chargerServiceFacade;
     }
