+++ src/main/java/com/example/happyDream/RestController/ChargerSwaggerRestController.java
@@ -2,6 +2,7 @@ package com.example.happyDream.RestController;
 
 import com.example.happyDream.DTO.ChargerDTO;
 import com.example.happyDream.DTO.ResponseDTO;
+import com.example.happyDream.Interface.ChargerSwagger;
 import com.example.happyDream.Service.ChargerServiceFacade;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
@@ -13,11 +14,11 @@ import java.util.*;
 @Slf4j
 @RestController //ResponseBody + Controller 어노테이션
 @RequestMapping(path = "/api/v1", produces = "application/json")
-public class ChargerRestController implements Charger{
+public class ChargerSwaggerRestController implements ChargerSwagger {
     private final ChargerServiceFacade chargerServiceFacade;
 
     @Autowired
-    public ChargerRestController(ChargerServiceFacade chargerServiceFacade) {
+    public ChargerSwaggerRestController(ChargerServiceFacade chargerServiceFacade) {
         this.chargerServiceFacade = chargerServiceFacade;
     }
